requirements.txt
django
drf-yasg
djangorestframework
djangorestframework-simplejwt
PyJWT
